use different transformation to include diacritic characters
Previously characters were transformed to u8 and b'A' got subtracted. In
non-optimized builds this caused range errors for characters outside of
the range 'A'-'Z', 'a'-'z'

Now the least significant 5 bits are taken which guarantees that no
exceptions can occur as the resulting numbers are allowed shifts for
1u32. There are collisions but they seem to be well distributed, see new
file for table.